Genomic Relationship Matrix construciton for coregenes
The Genomic  relationship  matrix  (G) uses  genotypic  data  to
estimate  the  fraction  of  total  DNA  that  two  individuals  share.
This is a measure of covariance between pairs of individuals, therefore
negatives values may exist because it calculates deviations from the
average. The genetic matrix (NxM dimensions) is defined as rows being
the individuals/strains (N) and columns as markers (M), missing data are
replaced with the mean frequency of the corresponding allele, which
avoids adding artefactual between-group differentiation.
The Genomic Relationship Matrix (G) was obtained by first centered the
genetic matrix (Z): subtracting the mean of the marker genotype and
standardized to a unit variance. After the normalization, the Genomic
Re-lationship Matrix was produced by computing the product: G = ZZ’/M,
where M is the total number of markers.